requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1
pandas>=2.1.0
openpyxl>=3.1.2
gunicorn>=21.2.0
